updated testcases and removed test_crud.py
