+++ request-api/makefile
@@ -11,10 +11,13 @@ black:
 flake8:
 	flake8 .
 
-test-coverage:: coverage-unit coverage-integration
+test-coverage:: coverage-unit coverage-integration coverage-acceptance
 
 coverage-unit:
 	pytest --cov=src tests/unit/
 
 coverage-integration:
 	pytest --cov=src --cov-append --cov-fail-under=90 tests/integration/
+
+coverage-acceptance:
+	pytest --cov=src --cov-append --cov-fail-under=90 tests/acceptance/
